requirements.txt
openai==1.61.1
langchain==0.3.19
langchain-openai==0.3.7
langchain-core==0.3.40
langchain-community==0.3.18
langchain-experimental==0.3.4
# langchain_chroma==0.2.2  # Commented out to avoid chromadb dependency issues
langchain-text-splitters==0.3.6
langsmith==0.3.11
# chromadb==0.6.3  # Replaced with faiss-cpu for vector operations
faiss-cpu==1.10.0
gradio==5.20.0
unstructured==0.16.23
google-search-results==2.4.2
tiktoken==0.9.0
docarray==0.40.0
matplotlib==3.10.0
plotly==5.24.1
numpy==1.26.4
scikit-learn==1.6.1
pandas==2.2.2
pillow~=11.3.0
PyYAML~=6.0.3
python-dotenv~=1.1.1
requests~=2.32.5